Crate monotone_crescendo

A library for calculating the least number of flips needed to make a binary string monotone increasing. Also contains utility functions for allocating webassembly memory and returning a pointer to it, as well as deallocating said memory. It also contains a function that can be called via WebAssembly, and thus Javascript.

Functions

alloc

Allocate memory in WebAssembly’s linear memory and return its offset

dealloc⚠

Deallocate memory in WebAssembly’s linear memory at the offset located at ptr

make_me_monotone_crescendo⚠

monotone_crescendo_prefix_sums

This is the official solution to LeetCode problem #926

monotone_crescendo_prefix_sums_without_redundant_zero

Essentially the same solution as monotone_crescendo_prefix_sums except without the redundant leading zero in the prefix sum array

monototone_crescendo